[refactor] 인증관련 파일 위치 변경
