+++ module-common/src/main/java/com/welcommu/modulecommon/token/TokenDto.java
@@ -1,4 +1,4 @@
-package com.welcommu.modulecommon.token.dto;
+package com.welcommu.modulecommon.token;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
